Move attr and slot_attr changeset validations to component.ex
